Duda de Poiseuille
Hello, how are you? I am a thesis student of the Master in Physics and my work is in SPH, blood flow simulation.
but before simulating I want to calibrate the code by Morris and Sigalotti's poiseuille profile which is very similar.
Then I have a strong doubt in the dualSPHysics library named Case_Poiseuille: where in the file do I declare the initial velocity? 
or that velocity is generated with parameters in an indirect way, thank you very much in advance.